src/requirements.txt
Pillow
luma.oled
RPi.GPIO
pyudev
psutil